refactor: consolidate react state tests
